+++ app/Http/Controllers/ViolationController.php
@@ -26,6 +26,8 @@ class ViolationController extends Controller
 
     public function destroy(Violation $violation)
     {
+        if ($violation->unitViolation)
+            return response()->json(['status' => 400, 'msg' => 'تم تسجيل مخالفات تحت مسمى هذه المخالفه من قبل لذلك لا يمكن حذفها']);
         $violation->delete();
         return response()->json(['status' => 200, 'msg' => 'تم حذف المخالفة']);
     }
